\name{aex}
\alias{aex}
\docType{data}
\title{
A Meta-Analysis for Evaluating the Effect of Aerobic Exercise on Visceral Adipose Tissue Content/Volume
}
\description{
This meta-analysis serves as an example to illustrate the function usage in the package \pkg{altmeta}.
}
\usage{data("aex")}
\format{
  A data frame containing 29 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{The observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{The within-study variance for each study.}
  }
}
\source{
Ismail I, Keating SE, Baker MK, and Johnson NA (2012). "A systematic review and meta-analysis of the effect of aerobic vs resistance exercise training on visceral fat." \emph{Obesity Reviews}, \bold{13}(1), 68--91.
}
\keyword{datasets}
