\name{dat.adams}
\alias{dat.adams}
\docType{data}
\title{
A Meta-Analysis on the Effects of Fluvastatin 20 mg/day on High-Density Lipoprotein
}
\description{
This meta-analysis serves as an example to illustrate the usage of the function \code{\link{meta.pen}}.
}
\usage{data("dat.ha")}
\format{
  A data frame containing 32 studies.
  \describe{
    \item{\code{y}}{the observed effect size (mean difference) for each study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Adams SP, Sekhon SS, Tsang M, Wright JM (2018). "Fluvastatin for lowering lipids." \emph{Cochrane Database of Systematic Reviews}, \bold{7}. Art.{\sspace}No.: CD012282. <\doi{10.1002/14651858.CD012282.pub2}>
}
\keyword{dataset}