\name{dat.aex}
\alias{dat.aex}
\docType{data}
\title{
A Meta-Analysis for Evaluating the Effect of Aerobic Exercise on Visceral Adipose Tissue Content/Volume
}
\description{
This meta-analysis serves as an example to illustrate function usage in the package \pkg{altmeta}.
}
\usage{data("dat.aex")}
\format{
  A data frame containing 29 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Ismail I, Keating SE, Baker MK, Johnson NA (2012). "A systematic review and meta-analysis of the effect of aerobic vs.{\sspace}resistance exercise training on visceral fat." \emph{Obesity Reviews}, \bold{13}(1), 68--91. <\doi{10.1111/j.1467-789X.2011.00931.x}>
}
\keyword{dataset}
