\name{dat.barlow}
\alias{dat.barlow}
\docType{data}
\title{
A Meta-Analysis on the Effect of Parent Training Programs vs. Control for Improving Parental Psychosocial Health Within 4 Weeks After Intervention
}
\description{
This dataset serves as an example of meta-analysis of standardized mean differences.
}
\usage{data("dat.barlow")}
\format{
  A data frame with 26 studies with the following 5 variables within each study.
  \describe{
    \item{\code{y}}{point esimates of standardized mean differences.}
    \item{\code{s2}}{sample variances of standardized mean differences.}
    \item{\code{n1}}{sample sizes in treatment group 1 (parent training programs).}
    \item{\code{n2}}{sample sizes in treatment group 2 (control).}
    \item{\code{n}}{total sample sizes.}
  }
}
\source{
Barlow J, Smailagic N, Huband N, Roloff V, Bennett C (2014). "Group-based parent training programmes for improving parental psychosocial health." \emph{Cochrane Database of Systematic Reviews}, \bold{5}, Art.{\sspace}No.: CD002020. <\doi{10.1002/14651858.CD002020.pub4}>
}
\keyword{dataset}
