\name{dat.beck17}
\alias{dat.beck17}
\docType{data}
\title{
A Meta-Analysis of Prevalence of Depression or Depressive Symptoms Among Medical Students
}
\description{
This dataset serves as an example of meta-analysis of proportions.
}
\usage{data("dat.beck17")}
\format{
  A data frame with 6 studies with the following 2 variables within each study.
  \describe{
    \item{\code{e}}{event counts of samples with depression or depressive symptoms.}
    \item{\code{n}}{sample sizes.}
  }
}
\details{
The original article by Rotenstein et al.{\sspace}(2016) stratified all extracted studies based on various screening instruments and cutoff scores. This dataset focuses on the meta-analysis of 6 studies with Beck Depression Inventory Score \eqn{\geq}17.
}
\source{
Rotenstein LS, Ramos MA, Torre M, Segal JB, Peluso MJ, Guille C, Sen S, Mata DA (2016). "Prevalence of depression, depressive symptoms, and suicidal ideation among medical students: a systematic review and meta-analysis." \emph{JAMA}, \bold{316}(21), 2214--2236. <\doi{10.1001/jama.2016.17324}>
}
\keyword{dataset}
