\name{dat.bellamy}
\alias{dat.bellamy}
\docType{data}
\title{
A Meta-Analysis on Type 2 Diabetes Mellitus After Gestational Diabetes
}
\description{
This meta-analysis serves as an example of meta-analysis with binary outcomes.
}
\usage{data("dat.bellamy")}
\format{
  A data frame containing 20 cohort studies with the following 4 variables.
  \describe{
    \item{\code{sid}}{study IDs.}
    \item{\code{tid}}{treatment/exposure IDs (0: non-exposure; 1: exposure).}
    \item{\code{e}}{event counts.}
    \item{\code{n}}{sample sizes.}
  }
}
\source{
Bellamy L, Casas JP, Hingorani AD, Williams D (2009). "Type 2 diabetes mellitus after gestational diabetes: a systematic review and meta-analysis." \emph{Lancet}, \bold{373}(9677), 1773--1779. <\doi{10.1016/S0140-6736(09)60731-5}>
}
\keyword{dataset}
