\name{dat.chor}
\alias{dat.chor}
\docType{data}
\title{
A Meta-Analysis of Proportions on Chorioamnionitis
}
\description{
This dataset serves as an example of meta-analysis of proportions.
}
\usage{data("dat.chor")}
\format{
  A data frame with 21 studies with the following 2 variables within each study.
  \describe{
    \item{\code{e}}{event counts of horioamnionitis.}
    \item{\code{n}}{sample sizes.}
  }
}
\source{
Woodd SL, Montoya A, Barreix M, Pi L, Calvert C, Rehman AM, Chou D, Campbell OMR (2019). "Incidence of maternal peripartum infection: a systematic review and meta-analysis." \emph{PLOS Medicine}, \bold{16}(12), e1002984. <\doi{10.1371/journal.pmed.1002984}>
}
\keyword{dataset}