\name{dat.ha}
\alias{dat.ha}
\docType{data}
\title{
A Meta-Analysis on the Effect of Placebo Interventions for All Clinical Conditions Regarding Patient-Reported Outcomes
}
\description{
This meta-analysis serves as an example to illustrate function usage in the package \pkg{altmeta}.
}
\usage{data("dat.ha")}
\format{
  A data frame containing 109 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Hrobjartsson A, Gotzsche PC (2010). "Placebo interventions for all clinical conditions." \emph{Cochrane Database of Systematic Reviews}, \bold{1}. Art.{\sspace}No.: CD003974. <\doi{10.1002/14651858.CD003974.pub3}>
}
\keyword{dataset}