\name{dat.hipfrac}
\alias{dat.hipfrac}
\docType{data}
\title{
A Meta-Analysis on the Magnitude and Duration of Excess Mortality After Hip Fracture Among Older Men
}
\description{
This meta-analysis serves as an example to illustrate function usage in the package \pkg{altmeta}.
}
\usage{data("dat.hipfrac")}
\format{
  A data frame containing 17 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Haentjens P, Magaziner J, Colon-Emeric CS, Vanderschueren D, Milisen K, Velkeniers B, Boonen S (2010). "Meta-analysis: excess mortality after hip fracture among older women and men". \emph{Annals of Internal Medicine}, \bold{152}(6), 380--390. <\doi{10.7326/0003-4819-152-6-201003160-00008}>
}
\keyword{dataset}