\name{dat.hughes}
\alias{dat.hughes}
\docType{data}
\title{
A Meta-Analysis on the Effects of Ovulation Suppression for Endometriosis in Subfertile Women
}
\description{
This meta-analysis serves as an example to illustrate the usage of the function \code{\link{metahet.hybrid}}.
}
\usage{data("dat.hughes")}
\format{
  A data frame containing 11 studies.
  \describe{
    \item{\code{y}}{the observed effect size (log odds ratio) for each study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Hughes E, Brown J, Collins JJ, et al. (2007). "Ovulation suppression for endometriosis for women with subfertility." \emph{Cochrane Database of Systematic Reviews}, \bold{7}. Art.{\sspace}No.: CD000155. <\doi{10.1002/14651858.CD000155.pub2}>
}
\keyword{dataset}