\name{dat.kaner}
\alias{dat.kaner}
\docType{data}
\title{
A Meta-Analysis on the Effect of Brief Alcohol Interventions vs. Control in Primary Care Populations
}
\description{
This dataset serves as an example of meta-analysis of risk differences.
}
\usage{data("dat.kaner")}
\format{
  A data frame with 13 studies with the following 7 variables within each study.
  \describe{
    \item{\code{y}}{point esimates of risk differences.}
    \item{\code{s2}}{sample variances of risk differences.}
    \item{\code{n1}}{sample sizes in treatment group 1 (brief alcohol interventions).}
    \item{\code{n2}}{sample sizes in treatment group 2 (control).}
    \item{\code{r1}}{event counts in treatment group 1.}
    \item{\code{r2}}{event counts in treatment group 2.}
    \item{\code{n}}{total sample sizes.}
  }
}
\source{
Kaner EF, Dickinson HO, Beyer FR, Campbell F, Schlesinger C, Heather N, Saunders JB, Burnand B, Pienaar ED (2007). "Effectiveness of brief alcohol interventions in primary care populations." \emph{Cochrane Database of Systematic Reviews}, \bold{2}, Art.{\sspace}No.: CD004148. <\doi{10.1002/14651858.CD004148.pub3}>
}
\keyword{dataset}