\name{dat.lcj}
\alias{dat.lcj}
\docType{data}
\title{
A Meta-Analysis on the Effect of Progressive Resistance Strength Training Exercise vs. Control
}
\description{
This meta-analysis serves as an example to illustrate function usage in the package \pkg{altmeta}.
}
\usage{data("dat.lcj")}
\format{
  A data frame containing 33 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Liu CJ, Latham NK (2009). "Progressive resistance strength training for improving physical function in older adults." \emph{Cochrane Database of Systematic Reviews}, \bold{3}. Art.{\sspace}No.: CD002759. <\doi{10.1002/14651858.CD002759.pub2}>
}
\keyword{dataset}