\name{dat.paige}
\alias{dat.paige}
\docType{data}
\title{
A Meta-Analysis on the Effectiveness of Spinal Manipulative Therapies (Other Than Sham)
}
\description{
This dataset serves as an example of meta-analysis of standardized mean differences.
}
\usage{data("dat.paige")}
\format{
  A data frame with 6 studies with the following 5 variables within each study.
  \describe{
    \item{\code{y}}{point esimates of standardized mean differences.}
    \item{\code{s2}}{sample variances of standardized mean differences.}
    \item{\code{n1}}{sample sizes in treatment group 1 (spinal manipulation).}
    \item{\code{n2}}{sample sizes in treatment group 2 (comparator).}
    \item{\code{n}}{total sample sizes.}
  }
}
\source{
Paige NM, Miake-Lye IM, Booth MS, Beroes JM, Mardian AS, Dougherty P, Branson R, Tang B, Morton SC, Shekelle PG (2017). "Association of spinal manipulative therapy with clinical benefit and harm for acute low back pain: systematic review and meta-analysis." \emph{JAMA}, \bold{317}(14), 1451--1460. <\doi{10.1001/jama.2017.3086}>
}
\keyword{dataset}