\name{dat.plourde}
\alias{dat.plourde}
\docType{data}
\title{
A Meta-Analysis for Comparing the Fluoroscopy Time in Percutaneous Coronary Intervention Between Radial and Femoral Accesses
}
\description{
This dataset serves as an example of meta-analysis of mean differences.
}
\usage{data("dat.plourde")}
\format{
  A data frame with 19 studies with the following 5 variables within each study.
  \describe{
    \item{\code{y}}{point esimates of mean differences.}
    \item{\code{s2}}{sample variances of mean differences.}
    \item{\code{n1}}{sample sizes in treatment group 1 (radial).}
    \item{\code{n2}}{sample sizes in treatment group 2 (femoral).}
    \item{\code{n}}{total sample sizes.}
  }
}
\source{
Plourde G, Pancholy SB, Nolan J, Jolly S, Rao SV, Amhed I, Bangalore S, Patel T, Dahm JB, Bertrand OF (2015). "Radiation exposure in relation to the arterial access site used for diagnostic coronary angiography and percutaneous coronary intervention: a systematic review and meta-analysis." \emph{Lancet}, \bold{386}(10009), 2192--2203. <\doi{10.1016/S0140-6736(15)00305-0}>
}
\keyword{dataset}