\name{dat.slf}
\alias{dat.slf}
\docType{data}
\title{
A Meta-Analysis on the Effect of Nicotine Gum for Smoking Cessation
}
\description{
This meta-analysis serves as an example to illustrate function usage in the package \pkg{altmeta}.
}
\usage{data("dat.slf")}
\format{
  A data frame containing 56 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{the within-study variance for each study.}
  }
}
\source{
Stead LF, Perera R, Bullen C, Mant D, Hartmann-Boyce J, Cahill K, Lancaster T (2012). "Nicotine replacement therapy for smoking cessation." \emph{Cochrane Database of Systematic Reviews}, \bold{11}. Art.{\sspace}No.: CD000146. <\doi{10.1002/14651858.CD000146.pub4}>
}
\keyword{dataset}