\name{dat.smith}
\alias{dat.smith}
\docType{data}
\title{
Meta-Analysis on the Diagnostic Accuracy of Ultrasound for Detecting Partial Thickness Rotator Cuff Tears
}
\description{
This meta-analysis serves as an example of meta-analyses of diagnostic tests.
}
\usage{data("dat.smith")}
\format{
  A data frame with 30 studies with the following 4 variables; each row represents a study.
  \describe{
    \item{\code{tp}}{counts of true positives.}
    \item{\code{fp}}{counts of false positives.}
    \item{\code{fn}}{counts of false negatives.}
    \item{\code{tn}}{counts of true negatives.}
  }
}
\source{
Smith TO, Back T, Toms AP, Hing CB (2011). "Diagnostic accuracy of ultrasound for rotator cuff tears in adults: a systematic review and meta-analysis." \emph{Clinical Radiology}, \bold{66}(11), 1036--1048. <\doi{10.1016/j.crad.2011.05.007}>
}
\keyword{dataset}