\name{hipfrac}
\alias{hipfrac}
\docType{data}
\title{
A Meta-Analysis for Investigating the Magnitude and Duration of Excess Mortality After Hip Fracture Among Older Men
}
\description{
This meta-analysis serves as an example to illustrate the function usage in the package \pkg{altmeta}.
}
\usage{data("hipfrac")}
\format{
  A data frame containing 17 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{The observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{The within-study variance for each study.}
  }
}
\source{
Haentjens P, Magaziner J, Colon-Emeric CS, Vanderschueren D, Milisen K, Velkeniers B, and Boonen S (2010). "Meta-analysis: excess mortality after hip fracture among older women and men". \emph{Annals of Internal Medicine}, \bold{152}(6), 380--390.
}
\keyword{datasets}
