\name{lcj}
\alias{lcj}
\docType{data}
\title{
A Meta-Analysis for Evaluating the Effect of Progressive Resistance Strength Training Exercise Versus Control
}
\description{
This meta-analysis serves as an example to illustrate the function usage in the package \pkg{altmeta}.
}
\usage{data("lcj")}
\format{
  A data frame containing 33 studies with the observed effect sizes and their within-study variances.
  \describe{
    \item{\code{y}}{The observed effect size for each collected study in the meta-analysis.}
    \item{\code{s2}}{The within-study variance for each study.}
  }
}
\source{
Liu CJ and Latham NK (2009). "Progressive resistance strength training for improving physical function in older adults." \emph{Cochrane Database of Systematic Reviews}, \bold{3}. Art No.: CD002759. DOI: 10.1002/14651858.CD002759.pub2.
}
\keyword{datasets}
